requirements.txt
tensorflow
streamlit
pandas
pillow
matplotlib
tqdm
numpy
transformers
torch
torchvision
